Fix REPL global variables
Globals that were available at the time of a function definition,
were not accessible inside calls to that function in the REPL,
because of the global namespace hacks that make the REPL work.

This works around the issue by copying all of the REPL globals into
the temp module and back. It's super inefficient, but
1) it's the REPL and
2) we already create an entire new module for each input line in
the REPL anyways